Moved states to game - vchadha
