+++ src/com/base/game/scenes/Event.java
@@ -1,0 +1,4 @@
+package com.base.game.scenes;
+
+public class Event {
+}
