+++ src/com/base/game/scenes/Scene.java
@@ -1,0 +1,8 @@
+package com.base.game.scenes;
+
+import java.util.List;
+
+public class Scene {
+    private List<Event> events;
+
+}
